ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
xgboost>=2.1.0
prophet>=1.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.28.0
